requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
cachetools>=5.3.2
youtube-transcript-api>=0.6.2 